Don't use 0 to disable Learning Mode. APPMODEL-86
Protobuf doesn't offer a reliable way to differentiate between a default
value being explicitly unset or skipped during serialisation. Doing
unset value detection may work incidentally but isn't guaranteed by
design.

Doorman learning mode duration was defaulted to a positive integer value
(of seconds) if it was unset. However, if it was set to 0 or a negative
integer, learning mode was disabled. The Unit Tests used the value 0 for
disabling learning mode. However, 0 being the default value for an
integer in Proto3, caused it to be skipped in serialisation and the call
site always defaulted in the learning mode duration instead of using 0.

The current change clarifies that only negative integer values would
disable learning mode and updates the unit tests to reflect the same.

Another way of handling this would be to update the proto3 definition to
specify the default duration in the spec.